Fix loop variable of inner loop in threads function
The inner loop in the the threads function used the same loop variable ("i") as the outer loop. As a consequence, at the end of the inner loop, i contains the last value of the inner loop iteration instead of the current outer loop i, and thus the thread id of a completely wrong message got updated. (However, this did, fortunately, not happen very often as the inner loop is executed only very very rarely.)
To fix this problem, rename the loop-variable of the inner loop to "j" instead.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>